Update sensor.py
As power meter provides a variation over the active power and both values should be combined in order to compute the power usage they should be obtained as close in time as possible. As they are in different memory sections it is not possible to retrieve them in a single modbus read so reading it in the following line seems the best workaround to prevent invalid power usage values.

I have other modifications in my fork like providing a prefix option in case you have two inverters of the same model you can prevent their entity name to conflict, i can make a PR for that if you think it may be useful. I have also a reduced attributes option to minimize modbus calls but that is based on my personal needs.

I am also considering passing the power usage and self consumption computations to attributes to the extension instead of making the active power and power meter active power substraction as sensor templates in home assistant but i am not convinced that adding derived values here is the way to go... but it will make users to not care abouut how they should be computed from the data returned by the inverter... what's your opinion?